Consolidate equipment changes
