+++ jsx/equipment/edit_equipment.jsx
