+++ jsx/equipment/equipment_form.jsx
