+++ jsx/equipment/equipment_schema.jsx
