+++ jsx/equipment/view_equipment.jsx
